+++ code/html/custom.js
@@ -34,7 +34,7 @@ function sensorName(id) {
         "DHT", "Dallas", "Emon Analog", "Emon ADC121", "Emon ADS1X15",
         "HLW8012", "V9261F", "ECH1560", "Analog", "Digital",
         "Events", "PMSX003", "BMX280", "MHZ19", "SI7021",
-        "SHT3X I2C"
+        "SHT3X I2C", "BH1750"
     ];
     if (1 <= id && id <= names.length) return names[id-1];
     return null;
@@ -44,9 +44,9 @@ function magnitudeType(type) {
     var types = [
         "Temperature", "Humidity", "Pressure",
         "Current", "Voltage", "Active Power", "Apparent Power",
-        "Reactive Power", "Energy", "Energy (delta)", "Power Factor",
+        "Reactive Power", "Power Factor", "Energy", "Energy (delta)",
         "Analog", "Digital", "Events",
-        "PM1.0", "PM2.5", "PM10", "CO2"
+        "PM1.0", "PM2.5", "PM10", "CO2", "Lux"
     ];
     if (1 <= type && type <= types.length) return types[type-1];
     return null;
@@ -55,7 +55,7 @@ function magnitudeType(type) {
 function magnitudeError(error) {
     var errors = [
         "OK", "Out of Range", "Warming Up", "Timeout", "Wrong ID",
-        "CRC Error", "I2C Error"
+        "CRC Error", "I2C Error", "GPIO Error"
     ];
     if (0 <= error && error < errors.length) return errors[error];
     return "Error " + error;